if_vtnet: Defer updating generated MAC address until attached
This improves VirtIO spec compliance because the guest driver
is not suppose to notify the device prior to setting the DRIVER_OK
status, which coudl happen with the VIRTIO_NET_F_CTRL_MAC_ADDR.
But the VIRTIO_NET_F_MAC feature should always be negotiated so
this is rare situation.

There are still other situations where we can notify without
DRIVER_OK set.