if_vtnet: Schedule Rx task if pending items when enabling interrupt
Prior to V1, the driver would enable interupts and then notify the
host that DRIVER_OK. Since for V1, DRIVER_OK needs to be set before
notifying the virtqueues, there may be items in the queues waiting
to be processed by the time interrupts are enabled.

This fixes a bug where the first Rx queue would appear stuck, only
being fixable by an interface down/up.